+++ package.json
@@ -42,6 +42,9 @@
     "setupFilesAfterEnv": [
       "<rootDir>src/setupTests.ts"
     ],
+    "moduleNameMapper": {
+      "\\.(css|less)$": "identity-obj-proxy"
+    },
     "testMatch": [
       "<rootDir>src/**/__tests__/*.{ts,tsx}"
     ],
@@ -65,6 +68,7 @@
     "babel-plugin-styled-components": "^1.10.6",
     "cssnano": "^4.1.10",
     "husky": "^3.0.0",
+    "identity-obj-proxy": "^3.0.0",
     "jest-dom": "^3.5.0",
     "jest-styled-components": "^7.0.0-beta.1",
     "prettier": "^1.18.2",
@@ -77,7 +81,7 @@
     "typescript": "^3.5.3"
   },
   "dependencies": {
-    "@tippy.js/react": "^2.2.2",
+    "@tippy.js/react": "^3.1.1",
     "lodash": "^4.17.15",
     "lodash-es": "^4.17.15",
     "styled-components": "^5.0.0-beta.8",
